alternative-project-setup Refactored and updated readmes
