+++ src/features/README.md
@@ -1,0 +1,8 @@
+# Features
+
+This directory is where you place the various features of your application.
+A feature is a fairly large piece of your application. In our initial state there is only the `question` directory, as that is our main.
+
+However, you could have another section for a landing page, or for the results page etc. It is up to you how you want to organise it, and I have found that when things start getting too big, I move parts to new `features` sub-directories.
+
+Basically, a feature is a logical part of the application, and is often seen as a section of the user interface.
