+++ src/features/question/README.md
@@ -1,0 +1,4 @@
+# Question feature
+
+Here is where all the code that is specific to question is put. So, not shared utilities or things that are global etc. e.g. data, constants, and the DOMelements.
+Here you will find the listeners, handlers, and views specific to making questions work.
