+++ src/features/question/questionListeners.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import { handleNextQuestion } from './questionHandlers.js';
+
+export const nextQuestion = () => {
+  handleNextQuestion();
+};
