+++ src/features/question/questionViews.js
@@ -1,0 +1,41 @@
+'use strict';
+
+import { NEXT_QUESTION_BUTTON_ID } from '../../shared/data/constants.js';
+import { nextQuestion } from './questionListeners.js';
+import { createDOMElement } from '../../shared/utils/DOMElement.js';
+
+export const createNextQuestionButtonElement = () => {
+  const buttonElement = createDOMElement('button', {
+    id: NEXT_QUESTION_BUTTON_ID,
+  });
+
+  buttonElement.innerText = 'Next question';
+  buttonElement.addEventListener('click', nextQuestion);
+
+  return buttonElement;
+};
+
+export const createQuestionElement = (question) => {
+  const container = createDOMElement('div');
+  const title = createDOMElement('h1');
+  title.innerText = question.text;
+  container.appendChild(title);
+
+  const answerContainer = createDOMElement('ol');
+
+  for (const answerKey in question.answers) {
+    const answer = createAnswerElement(question.answers[answerKey]);
+    answerContainer.appendChild(answer);
+  }
+
+  container.appendChild(answerContainer);
+
+  return container;
+};
+
+const createAnswerElement = (answerText) => {
+  const answerElement = createDOMElement('li');
+  answerElement.innerText = answerText;
+
+  return answerElement;
+};
