--- src/handlers/README.md
@@ -1,9 +1,0 @@
-# Handlers
-
-Handler function define user interactions. They will:
-
-- read user input from events and from the DOM
-- read and modify program data variables
-- process user data with logic functions
-- update the DOM to show changes to the user
-- log any important information for developers
