--- src/listeners/README.md
@@ -1,5 +1,0 @@
-# Listeners
-
-Event listeners are used to connect handler functions to user interactions with the DOM.
-
-You can use the same handler in many different listeners, or add more than one listeners to the same DOM element.
