+++ src/shared/README.md
@@ -1,0 +1,3 @@
+# Share directory
+
+This directory is where you put different things that get shared through out the application. Have a look through the sub-directories to get an idea of the things you would put here (usually in a subdirectory)
