+++ src/shared/components/README.md
@@ -1,0 +1,5 @@
+# Shared components
+
+If you find you are re-using a type of functionality, in different places in the application then you would store that here.
+
+We don't have any currently in the initial setup, but, for instance you might find that you are using a button over and over, or some part of your display such as a question. You could then put it in a library here, and call it wherever you need it in your application.
