+++ src/shared/data/README.md
@@ -1,0 +1,5 @@
+# Data
+
+This directory holds data that is shared throughout the application. For instance, Contstants that are global and should be accessible from throughout the application would go here. (see constants.js).
+
+Any other static data or starting models would also be stored here.
