+++ src/shared/utils/README.md
@@ -1,11 +1,4 @@
-# Logic
+# Utilities
 
-Logic functions are pure functions. They take primitives, objects or arrays as arguments and they return primitives, objects or arrays. Handlers will use logic functions to transform user input.
-
-Logic functions will _never_ ...
-
-- read from the DOM
-- write to the DOM
-- use events
-- use prompt/alert/confirm
-- use data that is not passed as a parameter
+This directory will contain libraries of utilities that are useful in the application. They usually have a general application.
+For instance, currently there is a library for dealing with the Document Object Model (DOM).
