--- src/shared/utils/getDOMElement.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const getDOMElement = (id) => {
-    return document.getElementById(id);
-}
-
-export default getDOMElement;
